Refactor acceptance/unit tests to use a shared file
