+++ docker-compose-dev.yml
@@ -6,7 +6,8 @@ services:
   solr-acceptance:
     build:
       context: ./solr
-    profiles: ["dev", "ci", "solr"]
+    pull_policy: build
+    profiles: ["acceptance", "dev", "solr"]
     ports:
       - 8983:8983
     command:
@@ -19,16 +20,24 @@ services:
       context: ./frontend
       args:
         - VOLTO_VERSION=${VOLTO_VERSION}
+    pull_policy: build
     platform: linux/amd64
     profiles: ["dev"]
     environment:
-      RAZZLE_API_PATH: http://localhost:55001/plone
-      RAZZLE_INTERNAL_API_PATH: http://backend:55001/plone
+      RAZZLE_API_PATH: http://localhost:8080/Plone
+      RAZZLE_INTERNAL_API_PATH: http://backend:8080/Plone
     depends_on:
       - backend
-      - backend-acceptance
     ports:
       - 3000:3000
+  frontend-acceptance:
+    <<: *frontend
+    profiles: ["acceptance"]
+    environment:
+      RAZZLE_API_PATH: http://localhost:55001/plone
+      RAZZLE_INTERNAL_API_PATH: http://backend-acceptance:55001/plone
+    depends_on:
+      - backend-acceptance
 
   backend: &backend
     build:
@@ -36,6 +45,7 @@ services:
       dockerfile: Dockerfile
       args:
         - PLONE_VERSION=${PLONE_VERSION}
+    pull_policy: build
     platform: linux/amd64
     profiles: ["dev"]
     environment:
@@ -47,32 +57,16 @@ services:
     ports:
       - 8080:8080
 
-  backend-acceptance: &backend-acceptance
+  backend-acceptance:
     <<: *backend
+    profiles: ["acceptance"]
     build:
       context: ./backend
       dockerfile: Dockerfile.acceptance
       args:
         - PLONE_VERSION=${PLONE_VERSION}
+    pull_policy: build
     depends_on:
       - solr-acceptance
     ports:
       - 55001:55001
-
-  # CI Services
-  frontend-ci:
-    <<: *frontend
-    image: kitconcept/kitconcept-solr-frontend:${BASE_TAG}
-    profiles: ["ci"]
-    platform: linux/amd64
-    depends_on:
-      - backend-ci
-      - backend-ci-acceptance
-  backend-ci:
-    <<: *backend
-    image: kitconcept/kitconcept-solr-backend:${BASE_TAG}
-    profiles: ["ci"]
-  backend-ci-acceptance:
-    <<: *backend-acceptance
-    image: kitconcept/kitconcept-solr-acceptance:${BASE_TAG}
-    profiles: ["ci"]
