--- docker-compose-solr.yml
@@ -1,14 +1,0 @@
----
-name: kitconcept-solr-solr
-
-services:
-
-  solr:
-    build:
-      context: solr/
-    ports:
-      - 8983:8983
-    command:
-      - solr-precreate
-      - plone
-      - /plone-config
